Create azuriteUtils.ts
